pyproject.toml
[project]
name = "urlshortener"
version = "0.1.0"
description = ""
authors = [
    {name = "jakramp",email = "jakramp96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=shortener.settings --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]